external/webkit/LayoutTests/platform/android/webgl-conformance/dumpAsText/conformance/glsl/misc/shader-with-quoted-error.frag-expected.txt

CONSOLE MESSAGE: line 24: *** Error loading shader '[object WebGLShader]':INVALID_VALUE
shader with error directive using quotes will fail

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

test: shader with error directive using quotes will fail
default vertex shader
test fragment shaderPASS shader with error directive using quotes will fail
PASS successfullyParsed is true

TEST COMPLETE
